Add Android Build Portable Script (abps)
Introduction:

As I was starting to use the Fish Shell, I've seen that it couldn't be used
with pure POSIX based scripts or even those hardcoded for bash. One of those
scripts is the Android environment setup script (envsetup.sh)

The main purpose of this script will be to accomodate those
who use the Fish Shell (or any not compatible shell for that matter) to at
least build an Android ROM.

By default, the number of threads used are the total available on the computer.
That said, the user can modify as he/she wishes by modyfing the "NB_CORES"
attribute.