Win10Hook.sh: Add dynamic binding for the onboard audio controller
There was always an issue whilst using PulseAudio to redirect sound
from the host. This issue was that it caused sound popping and
HTML5 videos that were getting out of sync.

Let's pass the entire chipset onboard audio to the VM to completely
fix the issue even though the audio on the host is lost during
that time.